Switched copilot response api to LA backend
