+++ apps/Standalone/src/designer/app/AzureLogicAppsDesigner/laDesigner.tsx
@@ -33,8 +33,6 @@ import {
   BaseApiManagementService,
   BaseAppServiceService,
   BaseChatbotService,
-  BaseCopilotWorkflowEditorService,
-  InitCopilotWorkflowEditorService,
   BaseExperimentationService,
   BaseUserPreferenceService,
   BaseFunctionService,
@@ -74,9 +72,6 @@ import {
   getMissingRoleDefinitions,
   roleQueryKeys,
   isAgentWorkflow,
-  setIsWorkflowDirty,
-  setFocusNode,
-  changePanelNode,
 } from '@microsoft/logic-apps-designer';
 import axios from 'axios';
 import isEqual from 'lodash.isequal';
@@ -522,26 +517,6 @@ const DesignerEditor = () => {
                   getUpdatedWorkflow={getUpdatedWorkflow}
                   openFeedbackPanel={() => openPanel('Azure Feedback Panel has been opened')}
                   closeChatBot={() => dispatch(setIsChatBotEnabled(false))}
-                  enableWorkflowEditing={true}
-                  autoApply={true}
-                  onWorkflowProposed={(newWorkflow) => {
-                    if (newWorkflow.parameters) {
-                      setCurrentParameters(newWorkflow.parameters as unknown as ParametersData);
-                    }
-                    setWorkflow({
-                      ...newWorkflow,
-                      id: guid(),
-                    });
-                    DesignerStore.dispatch(setIsWorkflowDirty(true));
-                  }}
-                  getNodeVisuals={(nodeId) => {
-                    const meta = DesignerStore.getState().operations.operationMetadata[nodeId];
-                    return meta ? { iconUri: meta.iconUri, brandColor: meta.brandColor } : undefined;
-                  }}
-                  onNodeClick={(nodeId) => {
-                    DesignerStore.dispatch(setFocusNode(nodeId));
-                    DesignerStore.dispatch(changePanelNode(nodeId));
-                  }}
                 />
               ) : null}
               <div
@@ -946,20 +921,6 @@ const getDesignerServices = (
     subscriptionId,
     location,
   });
-
-  // Initialize CopilotWorkflowEditorService if API key is configured
-  const copilotEditorApiKey = import.meta.env.VITE_COPILOT_EDITOR_API_KEY;
-  const copilotEditorEndpoint = import.meta.env.VITE_COPILOT_EDITOR_ENDPOINT;
-  if (copilotEditorApiKey && copilotEditorEndpoint) {
-    const copilotEditorService = new BaseCopilotWorkflowEditorService({
-      endpoint: copilotEditorEndpoint,
-      apiKey: copilotEditorApiKey,
-      model: import.meta.env.VITE_COPILOT_EDITOR_MODEL || undefined,
-      deploymentName: import.meta.env.VITE_COPILOT_EDITOR_DEPLOYMENT || undefined,
-      apiVersion: import.meta.env.VITE_COPILOT_EDITOR_API_VERSION || undefined,
-    });
-    InitCopilotWorkflowEditorService(copilotEditorService);
-  }
 
   const customCodeService = new StandardCustomCodeService({
     apiVersion: '2018-11-01',
